fix_4015 Added PerformanceDataMetricsFile to save metrics data
