+++ ohara-it/src/test/scala/oharastream/ohara/it/performance/BasicTestPerformance.scala
@@ -16,7 +16,6 @@
 
 package oharastream.ohara.it.performance
 
-import java.io.{File, FileWriter}
 import java.util.concurrent.atomic.LongAdder
 import java.util.concurrent.{Executors, TimeUnit}
 
@@ -51,7 +50,6 @@ import scala.util.control.Breaks.break
   */
 abstract class BasicTestPerformance extends WithRemoteWorkers {
   protected val log: Logger = Logger(classOf[BasicTestPerformance])
-  //private[this] val logDataInfoFileName = s"${CommonUtils.randomString(5)}.csv"
 
   private[this] val topicKey: TopicKey = TopicKey.of("benchmark", CommonUtils.randomString(5))
   protected val topicApi: TopicApi.Access =
@@ -79,13 +77,6 @@ abstract class BasicTestPerformance extends WithRemoteWorkers {
 
   @Rule
   override def timeout: Timeout = Timeout.seconds(wholeTimeout)
-
-  private[this] val reportOutputFolderKey = PerformanceTestingUtils.REPORT_OUTPUT_KEY
-  private[this] val reportOutputFolder: File = mkdir(
-    new File(
-      value(reportOutputFolderKey).getOrElse("/tmp/performance")
-    )
-  )
 
   private[this] val logMetersFrequencyKey               = PerformanceTestingUtils.LOG_METERS_FREQUENCY_KEY
   private[this] val logMetersFrequencyDefault: Duration = 5 seconds
@@ -139,105 +130,12 @@ abstract class BasicTestPerformance extends WithRemoteWorkers {
     }
   }
 
-  protected def mkdir(folder: File): File = {
-    if (!folder.exists() && !folder.mkdirs()) throw new AssertionError(s"failed to create folder on $folder")
-    if (folder.exists() && !folder.isDirectory) throw new AssertionError(s"$folder is not a folder")
-    folder
-  }
-
-  /**
-    * cache all historical meters. we always create an new file with all meters so we have to cache them.
-    */
-  private[this] val reportBuilders = mutable.Map[ConnectorKey, PerformanceReport.Builder]()
-
-  private[this] def connectorReports(): Seq[PerformanceReport] = {
-    val connectorInfos = result(connectorApi.list())
-    connectorInfos.map { info =>
-      val duration = (info.metrics.meters.flatMap(_.duration) :+ 0L).max / 1000
-      val builder  = reportBuilders.getOrElseUpdate(info.key, PerformanceReport.builder)
-      builder.connectorKey(info.key)
-      builder.className(info.className)
-      info.metrics.meters.foreach(
-        meter =>
-          builder
-            .record(duration, meter.name, meter.value)
-            .record(duration, s"${meter.name}(inPerSec)", meter.valueInPerSec.getOrElse(0.0f))
-      )
-      builder.build
-    }
-  }
-
-  private[this] def simpleName(className: String): String = {
-    val index = className.lastIndexOf(".")
-    if (index != -1) className.substring(index + 1)
-    else className
-  }
-
-  private[this] def path(className: String, fileName: String): File = {
-    new File(
-      mkdir(new File(mkdir(new File(reportOutputFolder, simpleName(className))), this.getClass.getSimpleName)),
-      fileName
+  protected[this] def rowData(): Row = {
+    Row.of(
+      (0 until 10).map(index => {
+        Cell.of(s"c$index", CommonUtils.randomString())
+      }): _*
     )
-  }
-
-  /*private[this] def logDataInfos(inputDataInfos: Seq[DataInfo]): Unit = {
-    val file = path("inputdata", logDataInfoFileName)
-    if (file.exists() && !file.delete()) throw new RuntimeException(s"failed to remove file:$file")
-    val fileWriter = new FileWriter(file)
-    try {
-      fileWriter.write("messageNumber,messageSize\n")
-      inputDataInfos.foreach(inputDataInfo => {
-        fileWriter.write(s"${inputDataInfo.messageNumber},${inputDataInfo.messageSize}\n")
-      })
-    } finally Releasable.close(fileWriter)
-  }*/
-
-  private[this] def logMeters(report: PerformanceReport): Unit = if (report.records.nonEmpty) {
-    // we have to fix the order of key-value
-    // if we generate line via map.keys and map.values, the order may be different ...
-    val headers = report.records.head._2.keySet.toList
-    report.records.values.foreach { items =>
-      headers.foreach(
-        name =>
-          if (!items.contains(name))
-            throw new RuntimeException(s"$name disappear?? current:${items.keySet.mkString(",")}")
-      )
-    }
-    val file = path(report.className, s"${report.key.group()}-${report.key.name()}.csv")
-    if (file.exists() && !file.delete()) throw new RuntimeException(s"failed to remove file:$file")
-    val fileWriter = new FileWriter(file)
-    try {
-      fileWriter.write("duration," + headers.map(s => s"""\"$s\"""").mkString(","))
-      fileWriter.write("\n")
-      report.records.foreach {
-        case (duration, item) =>
-          val line = s"$duration," + headers.map(header => f"${item(header)}%.3f").mkString(",")
-          fileWriter.write(line)
-          fileWriter.write("\n")
-      }
-    } finally Releasable.close(fileWriter)
-  }
-
-  protected def sleepUntilEnd(): Long = {
-    try {
-      val end = CommonUtils.current() + durationOfPerformance.toMillis
-      while (CommonUtils.current() <= end) {
-        fetchAllLogs()
-        TimeUnit.MILLISECONDS.sleep(logMetersFrequency.toMillis)
-      }
-    } finally {
-      beforeEndSleepUntil(fetchAllLogs())
-    }
-    durationOfPerformance.toMillis
-  }
-
-  /**
-    * invoked after all metrics of connectors are recorded.
-    * Noted: it is always invoked even if we fail to record reports
-    * @param reports the stuff we record
-    */
-  protected def afterRecodingReports(reports: Seq[PerformanceReport]): Unit = {
-    // nothing by default
   }
 
   /**
@@ -293,6 +191,66 @@ abstract class BasicTestPerformance extends WithRemoteWorkers {
     result(connectorApi.get(connectorKey))
   }
 
+  //------------------------------[metrics function]---------------------------//
+  private[this] val metricsFile = new PerformanceDataMetricsFile()
+
+  /**
+    * cache all historical meters. we always create an new file with all meters so we have to cache them.
+    */
+  private[this] val reportBuilders = mutable.Map[ConnectorKey, PerformanceReport.Builder]()
+
+  private[this] def connectorReports(): Seq[PerformanceReport] = {
+    val connectorInfos = result(connectorApi.list())
+    connectorInfos.map { info =>
+      val duration = (info.metrics.meters.flatMap(_.duration) :+ 0L).max / 1000
+      val builder  = reportBuilders.getOrElseUpdate(info.key, PerformanceReport.builder)
+      builder.connectorKey(info.key)
+      builder.className(info.className)
+      info.metrics.meters.foreach(
+        meter =>
+          builder
+            .record(duration, meter.name, meter.value)
+            .record(duration, s"${meter.name}(inPerSec)", meter.valueInPerSec.getOrElse(0.0f))
+      )
+      builder.build
+    }
+  }
+
+  /*private[this] def fetchAllLogs(): Seq[PerformanceReport] = {
+    val reports = connectorReports()
+    fetchConnectorMetrics(reports)
+    /*val inputDataInfos = inputDataMetrics()
+    if (inputDataInfos.nonEmpty) fetchDataInfoMetrics(inputDataInfos)*/
+    reports
+  }*/
+
+  private[this] def fetchConnectorMetrics(reports: Seq[PerformanceReport]): Unit = {
+    try reports.foreach(metricsFile.logMeters)
+    catch {
+      case e: Throwable =>
+        log.error("failed to log meters", e)
+    } finally afterRecodingReports(reports)
+  }
+
+  /*private[this] def fetchDataInfoMetrics(inputDataInfos: Seq[DataInfo]): Unit = {
+    try logDataInfos(inputDataInfos)
+    catch {
+      case e: Throwable =>
+        log.error("failed to log input data metrics", e)
+    }
+  }*/
+
+  /**
+    * invoked after all metrics of connectors are recorded.
+    * Noted: it is always invoked even if we fail to record reports
+    * @param reports the stuff we record
+    */
+  protected def afterRecodingReports(reports: Seq[PerformanceReport]): Unit = {
+    // nothing by default
+  }
+
+  //------------------------------[core functions]------------------------------//
+
   protected def produce(timeout: Duration): (TopicKey, Long, Long) = {
     val numberOfRowsToFlush = 2000
     val producer = Producer
@@ -344,12 +302,20 @@ abstract class BasicTestPerformance extends WithRemoteWorkers {
     (count.longValue, totalSizeInBytes.longValue)
   }
 
-  protected[this] def rowData(): Row = {
-    Row.of(
-      (0 until 10).map(index => {
-        Cell.of(s"c$index", CommonUtils.randomString())
-      }): _*
-    )
+  protected def sleepUntilEnd(): Long = {
+    try {
+      val end = CommonUtils.current() + durationOfPerformance.toMillis
+      while (CommonUtils.current() <= end) {
+        val reports = connectorReports()
+        fetchConnectorMetrics(reports)
+        TimeUnit.MILLISECONDS.sleep(logMetersFrequency.toMillis)
+      }
+    } finally {
+      val reports = connectorReports()
+      fetchConnectorMetrics(reports)
+      beforeEndSleepUntil(reports)
+    }
+    durationOfPerformance.toMillis
   }
 
   /**
@@ -360,8 +326,6 @@ abstract class BasicTestPerformance extends WithRemoteWorkers {
     * example delete data.
     */
   protected def afterStoppingConnectors(connectorInfos: Seq[ConnectorInfo], topicInfos: Seq[TopicInfo]): Unit = {}
-
-  //------------------------------[core functions]------------------------------//
 
   @After
   def record(): Unit = {
@@ -379,30 +343,6 @@ abstract class BasicTestPerformance extends WithRemoteWorkers {
     )
     afterStoppingConnectors(result(connectorApi.list()), result(topicApi.list()))
   }
-
-  private[this] def fetchAllLogs(): Seq[PerformanceReport] = {
-    val reports = connectorReports()
-    fetchConnectorMetrics(reports)
-    /*val inputDataInfos = inputDataMetrics()
-    if (inputDataInfos.nonEmpty) fetchDataInfoMetrics(inputDataInfos)*/
-    reports
-  }
-
-  private[this] def fetchConnectorMetrics(reports: Seq[PerformanceReport]): Unit = {
-    try reports.foreach(logMeters)
-    catch {
-      case e: Throwable =>
-        log.error("failed to log meters", e)
-    } finally afterRecodingReports(reports)
-  }
-
-  /*private[this] def fetchDataInfoMetrics(inputDataInfos: Seq[DataInfo]): Unit = {
-    try logDataInfos(inputDataInfos)
-    catch {
-      case e: Throwable =>
-        log.error("failed to log input data metrics", e)
-    }
-  }*/
 }
 
-final case class DataInfo(messageNumber: Long, messageSize: Long)
+final case class DataInfo(duration: Long, messageNumber: Long, messageSize: Long)
